Refactor:  유저인포dropdown컴포넌트 구현
