+++ webapp/src/components/Navigation/LoginNav.jsx
@@ -1,11 +1,10 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { Link, useNavigate } from 'react-router-dom';
-import Menu from 'components/Menu';
-import useModal from 'hooks/useModal';
+import { Link } from 'react-router-dom';
 import { HOME, MY_POST, NEW_POST, PROFILE, TEAM, MY_LIST, USER } from 'constant/route';
-import { deleteUserInfo } from 'service/auth';
+import useDropdown from 'hooks/useDropdown';
 import { Ul } from './style';
+import UserInfoDropdown from './UserInfoDropdown';
 
 LoginNav.propTypes = {
   userInfo: PropTypes.shape({
@@ -16,16 +15,11 @@ LoginNav.propTypes = {
 };
 
 export default function LoginNav({ userInfo }) {
-  const navigate = useNavigate();
+  const [parent, isDropdownOpen, shouldCloseDropdown, openDropdown, closeDropdown] = useDropdown();
   const { name, profileImg } = userInfo;
-  const [showModal, onCloseModal, openModal] = useModal();
-  const triggerLogOut = () => {
-    deleteUserInfo();
-    navigate('/');
-    window.location.reload();
-  };
+
   return (
-    <Ul>
+    <Ul ref={parent}>
       <li>
         <Link to={HOME}>Main</Link>
       </li>
@@ -38,26 +32,15 @@ export default function LoginNav({ userInfo }) {
       <li>
         <Link to={NEW_POST}>New Post</Link>
       </li>
-      <li onClick={openModal}>
+      <li onClick={openDropdown}>
         <img style={{ width: '30px' }} src={profileImg} alt="profile" />
         <span>{name}</span>
       </li>
-      <Menu style={{ right: 0, top: 80 }} show={showModal} onCloseModal={onCloseModal}>
-        <Ul IsflexDirectionColumn>
-          <li>
-            <Link to={MY_POST}>내 작성글</Link>
-          </li>
-          <li>
-            <Link to={MY_LIST}>내 관심글</Link>
-          </li>
-          <li>
-            <Link to={PROFILE}>프로필 설정</Link>
-          </li>
-          <li>
-            <button onClick={triggerLogOut}>로그아웃</button>
-          </li>
-        </Ul>
-      </Menu>
+      <UserInfoDropdown
+        isDropdownOpen={isDropdownOpen}
+        shouldCloseDropdown={shouldCloseDropdown}
+        closeDropdown={closeDropdown}
+      />
     </Ul>
   );
 }
